requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
PyMuPDF
openpyxl
python-dotenv
Pillow
requests
